Get sign up page to basic working state
I haven't verified the data is actually being saved into the database, but the sign up page no longer throws any errors and all the appropriate methods seem to be getting called.
I've added a bunch of annotations to various classes that were missing them, using @Autowired should be a lot more forgiving now. There were also some annonations which were attached to the wrong classes, so I've moved those to where it seems like they should be.

Upon researching the mongo connection annotations and setup a bit more, it seems like the currently existing classes called '__Repository' are meant to be the direct interaction points with the database. I created a new class 'UsersDatabase' according to templates found online, but from the look of things that can probably be combined with 'NewUserRepository' at some point.

I also haven't added any form of confirmation to the front end. Currently the user will *always* attempt to be created, and the success page will *always* be shown. This should be fixed by changing the add user method on the controller to return some form of indication about whether registration was successful or not.